:recycle: (901): move S3 audio fetch into infrastructure/s3.py
